factor out models/validators and populate publication graph with user list
