+++ designsafe/apps/projects_v2/schema_models/hybrid_sim.py
@@ -1,15 +1,17 @@
 """Pydantic schema models for Hybrid Simulation entities"""
 from typing import Annotated, Optional
 from pydantic import BeforeValidator, Field, model_validator
-from designsafe.apps.projects_v2.schema_models.base import (
-    MetadataModel,
+from designsafe.apps.projects_v2.schema_models._field_models import MetadataModel
+from designsafe.apps.projects_v2.schema_models._field_models import (
     AssociatedProject,
-    ReferencedWork,
-    ProjectUser,
-    FileTag,
-    FileObj,
-    Ref,
     DropdownValue,
+    FileObj,
+    FileTag,
+    ProjectUser,
+    Ref,
+    ReferencedWork,
+)
+from designsafe.apps.projects_v2.schema_models._field_transforms import (
     handle_array_of_none,
     handle_legacy_authors,
     handle_dropdown_value,
