+++ designsafe/apps/projects_v2/schema_models/simulation.py
@@ -2,17 +2,19 @@
 
 from typing import Optional, Annotated
 from pydantic import BeforeValidator, ConfigDict, Field, model_validator
-from designsafe.apps.projects_v2.schema_models.base import (
-    MetadataModel,
+from designsafe.apps.projects_v2.schema_models._field_models import MetadataModel
+from designsafe.apps.projects_v2.schema_models._field_models import (
     AssociatedProject,
-    ReferencedWork,
-    ProjectUser,
-    FileTag,
-    Ref,
     DropdownValue,
     FileObj,
-    handle_legacy_authors,
+    FileTag,
+    ProjectUser,
+    Ref,
+    ReferencedWork,
+)
+from designsafe.apps.projects_v2.schema_models._field_transforms import (
     handle_array_of_none,
+    handle_legacy_authors,
     handle_dropdown_value,
 )
 from designsafe.apps.projects_v2.constants import SIMULATION_TYPES
